requirements.txt
docker
configparser
ansicolors
pyinstaller
